Move pkg/test/shell under pkg/test/upgrade/shell
