+++ test/upgrade/shell/fail-example.sh
